[CABMAN] Move CCFDATAStorage to a separate header
