+++ sdk/tools/cabman/cabinet.h
@@ -308,29 +308,6 @@ public:
 
 /* Classes */
 
-#ifndef CAB_READ_ONLY
-
-class CCFDATAStorage
-{
-public:
-    /* Default constructor */
-    CCFDATAStorage();
-    /* Default destructor */
-    virtual ~CCFDATAStorage();
-    ULONG Create();
-    ULONG Destroy();
-    ULONG Truncate();
-    ULONG Position();
-    ULONG Seek(LONG Position);
-    ULONG ReadBlock(PCFDATA Data, void* Buffer, PULONG BytesRead);
-    ULONG WriteBlock(PCFDATA Data, void* Buffer, PULONG BytesWritten);
-private:
-    char FullName[PATH_MAX];
-    FILE* FileHandle;
-};
-
-#endif /* CAB_READ_ONLY */
-
 class CCabinet
 {
 public:
@@ -508,7 +485,7 @@ private:
     bool CreateNewDisk;
     bool CreateNewFolder;
 
-    CCFDATAStorage *ScratchFile;
+    class CCFDATAStorage *ScratchFile;
     FILE* SourceFile;
     bool ContinueFile;
     ULONG TotalBytesLeft;
